tensorflow2_django_api_asgi/requirements.txt
django
djangorestframework
django-cors-headers
uvicorn
pillow